third_party/WebKit/LayoutTests/fast/canvas/OffscreenCanvas-getContext-expected.txt

Tests basic functionalities of offscreenCanvas.getContext on the main thread

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS aCanvas.width is 40
PASS aCanvas.height is 60
PASS aCanvas.width is 110
PASS aCanvas.height is 90
PASS aCanvas.width is 0
PASS ctx = aCanvas.getContext('2d') did not throw exception.
PASS ctx is an instance of OffscreenCanvasRenderingContext2D
PASS ctx2 is null
PASS ctx3 is non-null.
PASS ctx3 == ctx is true
PASS ctx4 is an instance of WebGLRenderingContext
PASS successfullyParsed is true

TEST COMPLETE
